requirements.txt
asgiref==3.4.1
autocorrect==2.5.0
certifi==2021.5.30
chardet==4.0.0
collection==0.1.6
Django==3.2.6
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
docx==0.2.4
idna==2.10
language-tool-python==2.5.4
lxml==4.6.3
Markdown==3.3.4
Pillow==8.2.0
psycopg2-binary==2.9.1
python-docx==0.8.11
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
tqdm==4.61.1
urllib3==1.26.5
